Add files via upload
https://blueteamlabs.online/home/challenge/network-analysis-ransomware-3dd520c7ec
Scenario
ABC Industries worked day and night for a month to prepare a tender document for a prestigious project that would secure the company’s financial future. The company was hit by ransomware, believed to be conducted by a competitor, and the final version of the tender document was encrypted. Right now they are in need of an expert who can decrypt this critical document. All we have is the network traffic, the ransom note, and the encrypted ender document. Do your thing Defender!​

Challenge Submission
Q:What is the operating system of the host from which the network traffic was captured? (Look at Capture File Properties, copy the details exactly) (3 points)
A:32-bit Windows & Service PAck 1, build 7601

Q:What is the full URL from which the ransomware executable was downloaded? (3 points)
A: http://10.0.2.15:8000/safecrypt.exe

Q:Name the ransomware executable file? (2 points)
A: safecrypt.exe

Q:What is the MD5 hash of the ransomware? (2 points)
A:4a1d88603b1007825a9c6b36d1e5de44

Q:What is the name of the ransomware? (2 points)
A: Teslacrypt

Q:What is the encryption algorithm used by the ransomware, according to the ransom note? (2 points)
A: RSA-4096

Q:What is the domain beginning with ‘d’ that is related to ransomware traffic? (3 points)
A: dunyamuzelerimuzesi.com

Q:Decrypt the Tender document and submit the flag (3 points)
A: BTLO-T3nd3r-Fl@g